refactor and move gpu replay tracer
